move code from main into separate files
existing functions:
    findoldlogs
    handlerequest
    handleresponse
    outputinfo
    parse_program_args

new functions:
    find_namespaces
    output_program_args
    read_last_log
    summary_log_setup

also moved include logging.h to consts.h
