+++ src/rsrc_mgr/resourcemanager.c
@@ -59,52 +59,28 @@ GNU licenses can be found at http://www.gnu.org/licenses/.
 
 #define RMAN_USE_MPI
 
-#include "marfs_auto_config.h"
-#ifdef DEBUG_RM
-#define DEBUG DEBUG_RM
-#elif (defined DEBUG_ALL)
-#define DEBUG DEBUG_ALL
-#endif
-#define LOG_PREFIX "resourcemanager"
-#include "logging/logging.h"
-
-#include "config/config.h"
-#include "rsrc_mgr/resourcethreads.h"
-#include "rsrc_mgr/rmanstate.h"
-#include "rsrc_mgr/work.h"
-#include "rsrc_mgr/loginfo.h"
-
 #include <dirent.h>
 #include <mpi.h>
 #include <stdio.h>
 #include <sys/time.h>
 #include <sys/types.h>
 
-//   -------------   INTERNAL DEFINITIONS    -------------
+// specifically needed for this file
+#include "config/config.h"
+#include "rsrc_mgr/consts.h"
+#include "rsrc_mgr/loginfo.h"
+#include "rsrc_mgr/work.h"
 
-#define GC_THRESH 604800  // Age of deleted files before they are Garbage Collected
-                          // Default to 7 days ago
-#define RB_L_THRESH  600  // Age of files before they are rebuilt (based on location)
-                          // Default to 10 minutes ago
-#define RB_M_THRESH  120  // Age of files before they are rebuilt (based on marker)
-                          // Default to 2 minutes ago
-#define RP_THRESH 259200  // Age of files before they are repacked
-                          // Default to 3 days ago
-#define CL_THRESH  86400  // Age of intermediate state files before they are cleaned up (failed repacks, old logs, etc.)
-                          // Default to 1 day ago
-
-#define INACTIVE_RUN_SKIP_THRESH 60 // Age of seemingly inactive (no summary file) rman logdirs before they are skipped
-                                    // Default to 1 minute ago
-
-#define DEFAULT_LOG_ROOT "/var/log/marfs-rman"
-#define MODIFY_ITERATION_PARENT "RMAN-MODIFY-RUNS"
-#define RECORD_ITERATION_PARENT "RMAN-RECORD-RUNS"
-#define SUMMARY_FILENAME "summary.log"
-#define ERROR_LOG_PREFIX "ERRORS-"
-#define ITERATION_ARGS_FILE "PROGRAM-ARGUMENTS"
-#define OLDLOG_PREALLOC 16  // pre-allocate space for 16 logfiles in the oldlogs hash table (double from there, as needed)
-
-//   -------------   HELPER FUNCTIONS    -------------
+// rmanstate initialization
+// TODO: merge into rmanstate
+#include "rsrc_mgr/find_namespaces.h"
+#include "rsrc_mgr/findoldlogs.h"
+#include "rsrc_mgr/output_program_args.h"
+#include "rsrc_mgr/outputinfo.h"
+#include "rsrc_mgr/parse_program_args.h"
+#include "rsrc_mgr/read_last_log.h"
+#include "rsrc_mgr/rmanstate.h"
+#include "rsrc_mgr/summary_log_setup.h"
 
 static void print_usage_info() {
    printf("\n"
@@ -156,1350 +132,6 @@ static void print_usage_info() {
            "\n",
            DEFAULT_LOG_ROOT);
 }
-
-static int error_only_filter(const opinfo* op) {
-    return !op->errval;
-}
-
-static void outputinfo(FILE* output, marfs_ns* ns, streamwalker_report* report , operation_summary* summary) {
-   char userout = 0;
-   if (output == stdout) { userout = 1; } // limit info output directly to the user
-   fprintf(output, "Namespace \"%s\"%s --\n", ns->idstr, (userout) ? " Totals" : " Incremental Values");
-   fprintf(output, "   Walk Report --\n");
-   if (!userout || report->fileusage)   { fprintf(output, "      File Usage = %zu\n", report->fileusage); }
-   size_t bytetrans = report->byteusage;
-   char remainder[6] = {0};
-   char* unit = "B";
-   while (bytetrans > 1024) {
-      if (*unit == 'B') { unit = "KiB"; }
-      else if (*unit == 'K') { unit = "MiB"; }
-      else if (*unit == 'M') { unit = "GiB"; }
-      else if (*unit == 'G') { unit = "TiB"; }
-      else if (*unit == 'T') { unit = "PiB"; }
-      else { break; }
-      if (bytetrans % 1024) {
-         snprintf(remainder, 6, ".%.3zu", (((bytetrans % 1024) * 1000) + 512) / 1024);
-      } else { *remainder = '\0'; }
-      bytetrans /= 1024;
-   }
-   if (!userout)   { fprintf(output, "      Byte Usage = %zu\n", report->byteusage); }
-   else if (report->byteusage) {
-      fprintf(output, "      Byte Usage = %zu%s%s\n", bytetrans, remainder, unit);
-   }
-   if (!userout || report->filecount)   { fprintf(output, "      File Count = %zu\n", report->filecount); }
-   if (!userout || report->objcount)    { fprintf(output, "      Object Count = %zu\n", report->objcount); }
-   bytetrans = report->bytecount;
-   *remainder = '\0';
-   unit = "B";
-   while (bytetrans > 1024) {
-      if (*unit == 'B') { unit = "KiB"; }
-      else if (*unit == 'K') { unit = "MiB"; }
-      else if (*unit == 'M') { unit = "GiB"; }
-      else if (*unit == 'G') { unit = "TiB"; }
-      else if (*unit == 'T') { unit = "PiB"; }
-      else { break; }
-      if (bytetrans % 1024) {
-         snprintf(remainder, 6, ".%.3zu", (((bytetrans % 1024) * 1000) + 512) / 1024);
-      } else { *remainder = '\0'; }
-      bytetrans /= 1024;
-   }
-   if (!userout)   { fprintf(output, "      Byte Count = %zu\n", report->bytecount); }
-   else if (report->bytecount) {
-      fprintf(output, "      Byte Count = %zu%s%s\n", bytetrans, remainder, unit);
-   }
-   if (!userout || report->streamcount) { fprintf(output, "      Stream Count = %zu\n", report->streamcount); }
-   if (!userout || report->delobjs)     { fprintf(output, "      Object Deletion Candidates = %zu\n", report->delobjs); }
-   if (!userout || report->delfiles)    { fprintf(output, "      File Deletion Candidates = %zu\n", report->delfiles); }
-   if (!userout || report->delstreams)  { fprintf(output, "      Stream Deletion Candidates = %zu\n", report->delstreams); }
-   if (!userout || report->volfiles)    { fprintf(output, "      Volatile File Count = %zu\n", report->volfiles); }
-   if (!userout || report->rpckfiles)   { fprintf(output, "      File Repack Candidates = %zu\n", report->rpckfiles); }
-   if (!userout || report->rpckbytes)   { fprintf(output, "      Repack Candidate Bytes = %zu\n", report->rpckbytes); }
-   if (!userout || report->rbldobjs)    { fprintf(output, "      Object Rebuild Candidates = %zu\n", report->rbldobjs); }
-   if (!userout || report->rbldbytes)   { fprintf(output, "      Rebuild Candidate Bytes = %zu\n", report->rbldbytes); }
-   fprintf(output, "   Operation Summary --\n");
-   if (!userout || summary->deletion_object_count) {
-      fprintf(output, "      Object Deletion Count = %zu (%zu Failures)\n",
-               summary->deletion_object_count, summary->deletion_object_failures);
-   }
-   if (!userout || summary->deletion_reference_count) {
-      fprintf(output, "      Reference Deletion Count = %zu (%zu Failures)\n",
-               summary->deletion_reference_count, summary->deletion_reference_failures);
-   }
-   if (!userout || summary->rebuild_count) {
-      fprintf(output, "      Object Rebuild Count = %zu (%zu Failures)\n",
-               summary->rebuild_count, summary->rebuild_failures);
-   }
-   if (!userout || summary->repack_count) {
-      fprintf(output, "      File Repack Count = %zu (%zu Failures)\n",
-               summary->repack_count, summary->repack_failures);
-   }
-   fprintf(output, "\n");
-   fflush(output);
-   return;
-}
-
-static int output_program_args(rmanstate* rman) {
-   // start with marfs config version (config changes could seriously break an attempt to re-execute this later)
-   if (fprintf(rman->summarylog, "%s\n", rman->config->version) < 1) {
-      fprintf(stderr, "ERROR: Failed to output config version to summary log\n");
-      return -1;
-   }
-   // output operation types and threshold values
-   if (rman->gstate.dryrun  &&  fprintf(rman->summarylog, "DRY-RUN\n") < 1) {
-      fprintf(stderr, "ERROR: Failed to output dry-run flag to summary log\n");
-      return -1;
-   }
-   if (rman->quotas  &&  fprintf(rman->summarylog, "QUOTAS\n") < 1) {
-      fprintf(stderr, "ERROR: Failed to output quota flag to summary log\n");
-      return -1;
-   }
-   if (rman->gstate.thresh.gcthreshold  &&
-        fprintf(rman->summarylog, "GC=%llu\n", (unsigned long long) rman->gstate.thresh.gcthreshold) < 1) {
-      fprintf(stderr, "ERROR: Failed to output GC threshold to summary log\n");
-      return -1;
-   }
-   if (rman->gstate.thresh.repackthreshold  &&
-        fprintf(rman->summarylog, "REPACK=%llu\n", (unsigned long long) rman->gstate.thresh.repackthreshold) < 1) {
-      fprintf(stderr, "ERROR: Failed to output REPACK threshold to summary log\n");
-      return -1;
-   }
-   if (rman->gstate.thresh.rebuildthreshold  &&
-        fprintf(rman->summarylog, "REBUILD=%llu\n", (unsigned long long) rman->gstate.thresh.rebuildthreshold) < 1) {
-      fprintf(stderr, "ERROR: Failed to output REBUILD threshold to summary log\n");
-      return -1;
-   }
-   if (rman->gstate.lbrebuild  &&
-        (
-         fprintf(rman->summarylog, "REBUILD-LOCATION-POD=%d\n", rman->gstate.rebuildloc.pod) < 1  ||
-         fprintf(rman->summarylog, "REBUILD-LOCATION-CAP=%d\n", rman->gstate.rebuildloc.cap) < 1  ||
-         fprintf(rman->summarylog, "REBUILD-LOCATION-SCATTER=%d\n", rman->gstate.rebuildloc.scatter) < 1
-       )) {
-      fprintf(stderr, "ERROR: Failed to output REBUILD location to summary log\n");
-      return -1;
-   }
-   if (fprintf(rman->summarylog, "\n") < 1) {
-      fprintf(stderr, "ERROR: Failed to output header separator summary log\n");
-      return -1;
-   }
-   return 0;
-}
-
-static int parse_program_args(rmanstate* rman, FILE* inputsummary) {
-   // parse in a verify the config version
-   char* readline = NULL;
-   size_t linealloc = 0;
-   ssize_t linelen = getline(&readline, &linealloc, inputsummary);
-   if (linelen < 2) {
-      fprintf(stderr, "WARNING: Failed to read MarFS config version string from previous run's summary log\n");
-      return 1;
-   }
-   linelen--; // decrement length, to exclude newline
-   if (strncmp(readline, rman->config->version, linelen)) {
-      fprintf(stderr, "WARNING: Previous run is associated with a different MarFS config version: \"%s\"\n", readline);
-      free(readline);
-      return 1;
-   }
-   // parse in operation threshold values and set our state to match
-   while ((linelen = getline(&readline, &linealloc, inputsummary)) > 1) {
-      linelen--; // decrement length, to exclude newline
-      if (strncmp(readline, "QUOTAS", linelen) == 0) {
-         rman->quotas = 1;
-      }
-      else if (strncmp(readline, "DRY-RUN", linelen) == 0) {
-         rman->gstate.dryrun = 1;
-      }
-      else {
-         // look for an '=' char
-         char* parse = readline;
-         while (*parse != '=') {
-            if (*parse == '\0') {
-               fprintf(stderr, "ERROR: Failed to parse previous run's operation info (\"%s\")\n", readline);
-               free(readline);
-               return -1;
-            }
-            parse++;
-         }
-         *parse = '\0';
-         parse++;
-         // parse in the numeric value
-         char* endptr = NULL;
-         unsigned long long parseval = strtoull(parse, &endptr, 10);
-         if (endptr == NULL  ||  *endptr != '\n'  ||  parseval == ULLONG_MAX) {
-            fprintf(stderr, "ERROR: Failed to parse previous run's \"%s\" operation threshold: \"%s\"\n", readline, parse);
-            free(readline);
-            return -1;
-         }
-         // populate the appropriate value, based on string header
-         if (strcmp(readline, "GC") == 0) {
-            rman->gstate.thresh.gcthreshold = (time_t)parseval;
-         }
-         else if (strcmp(readline, "REPACK") == 0) {
-            rman->gstate.thresh.repackthreshold = (time_t)parseval;
-         }
-         else if (strcmp(readline, "REBUILD") == 0) {
-            rman->gstate.thresh.rebuildthreshold = (time_t)parseval;
-         }
-         else if (strcmp(readline, "REBUILD-LOCATION-POD") == 0) {
-            rman->gstate.rebuildloc.pod = (int)parseval;
-            rman->gstate.lbrebuild = 1;
-         }
-         else if (strcmp(readline, "REBUILD-LOCATION-CAP") == 0) {
-            rman->gstate.rebuildloc.cap = (int)parseval;
-            rman->gstate.lbrebuild = 1;
-         }
-         else if (strcmp(readline, "REBUILD-LOCATION-SCATTER") == 0) {
-            rman->gstate.rebuildloc.scatter = (int)parseval;
-            rman->gstate.lbrebuild = 1;
-         }
-         else {
-            fprintf(stderr, "ERROR: Encountered unrecognized operation type in log of previous run: \"%s\"\n", readline);
-            free(readline);
-            return -1;
-         }
-      }
-   }
-   if (linelen < 1) {
-      fprintf(stderr, "ERROR: Failed to read operation info from previous run's logfile\n");
-      return -1;
-   }
-   if (readline) { free(readline); } // done with this string allocation
-   return 0;
-}
-
-/**
- * Configure the current rank to target the given NS (configures state and launches worker threads)
- * @param rmanstate* rman : State struct for the current rank
- * @param marfs_ns* ns : New NS to target
- * @param workresponse* response : Response to populate with error description, on failure
- * @return int : Zero on success, -1 on failure
- */
-static int setranktgt(rmanstate* rman, marfs_ns* ns, workresponse* response) {
-   // update position value
-   if (config_establishposition(&rman->gstate.pos, rman->config)) {
-      LOG(LOG_ERR, "Failed to establish a root NS position\n");
-      snprintf(response->errorstr, MAX_ERROR_BUFFER, "Failed to establish a root NS position");
-      return -1;
-   }
-   char* tmpnspath = NULL;
-   if (config_nsinfo(ns->idstr, NULL, &tmpnspath)) {
-      LOG(LOG_ERR, "Failed to identify NS path of NS \"%s\"\n", ns->idstr);
-      snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                "Failed to identify NS path of NS \"%s\"", ns->idstr);
-      return -1;
-   }
-   char* nspath = strdup(tmpnspath+1); // strip off leading '/', to get a relative NS path
-   free(tmpnspath);
-   if (config_traverse(rman->config, &rman->gstate.pos, &nspath, 0)) {
-      LOG(LOG_ERR, "Failed to traverse config to new NS path: \"%s\"\n", nspath);
-      snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                "Failed to traverse config to new NS path: \"%s\"", nspath);
-      free(nspath);
-      return -1;
-   }
-   free(nspath);
-   if (rman->gstate.pos.ctxt == NULL  &&  config_fortifyposition(&rman->gstate.pos)) {
-      LOG(LOG_ERR, "Failed to fortify position for new NS: \"%s\"\n", ns->idstr);
-      snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                "Failed to fortify position for new NS: \"%s\"", ns->idstr);
-      config_abandonposition(&rman->gstate.pos);
-      return -1;
-   }
-   // update our resource input struct
-   if (resourceinput_init(&rman->gstate.rinput, &rman->gstate.pos, rman->gstate.numprodthreads)) {
-      LOG(LOG_ERR, "Failed to initialize resourceinput for new NS target: \"%s\"\n", ns->idstr);
-      snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                "Failed to initialize resourceinput for new NS target: \"%s\"", ns->idstr);
-      config_abandonposition(&rman->gstate.pos);
-      return -1;
-   }
-   // update our output resource log
-   char* outlogpath = resourcelog_genlogpath(1, rman->logroot, rman->iteration, ns, rman->ranknum);
-   if (outlogpath == NULL) {
-      LOG(LOG_ERR, "Failed to identify output logfile path of rank %zu for NS \"%s\"\n",
-                    rman->ranknum, ns->idstr);
-      snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                "Failed to identify output logfile path of rank %zu for NS \"%s\"",
-                rman->ranknum, ns->idstr);
-      resourceinput_purge(&rman->gstate.rinput);
-      config_abandonposition(&rman->gstate.pos);
-      return -1;
-   }
-   if (resourcelog_init(&rman->gstate.rlog, outlogpath,
-                          (rman->gstate.dryrun) ? RESOURCE_RECORD_LOG : RESOURCE_MODIFY_LOG, ns)) {
-      LOG(LOG_ERR, "Failed to initialize output logfile: \"%s\"\n", outlogpath);
-      snprintf(response->errorstr, MAX_ERROR_BUFFER, "Failed to initialize output logfile: \"%s\"", outlogpath);
-      free(outlogpath);
-      resourceinput_purge(&rman->gstate.rinput);
-      config_abandonposition(&rman->gstate.pos);
-      return -1;
-   }
-   free(outlogpath);
-   // update our repack streamer
-   if ((rman->gstate.rpst = repackstreamer_init()) == NULL) {
-      LOG(LOG_ERR, "Failed to initialize repack streamer\n");
-      snprintf(response->errorstr, MAX_ERROR_BUFFER, "Failed to initialize repack streamer");
-      resourcelog_term(&rman->gstate.rlog, NULL, 1);
-      resourceinput_purge(&rman->gstate.rinput);
-      config_abandonposition(&rman->gstate.pos);
-      return -1;
-   }
-   // kick off our worker threads
-   TQ_Init_Opts tqopts = {
-      .log_prefix = "RManWorker",
-      .init_flags = 0,
-      .max_qdepth = rman->gstate.numprodthreads + rman->gstate.numconsthreads,
-      .global_state = &rman->gstate,
-      .num_threads = rman->gstate.numprodthreads + rman->gstate.numconsthreads,
-      .num_prod_threads = rman->gstate.numprodthreads,
-      .thread_init_func = rthread_init_func,
-      .thread_consumer_func = rthread_consumer_func,
-      .thread_producer_func = rthread_producer_func,
-      .thread_pause_func = NULL,
-      .thread_resume_func = NULL,
-      .thread_term_func = rthread_term_func
-   };
-   if ((rman->tq = tq_init(&tqopts)) == NULL) {
-      LOG(LOG_ERR, "Failed to start ThreadQueue for NS \"%s\"\n", ns->idstr);
-      snprintf(response->errorstr, MAX_ERROR_BUFFER, "Failed to start ThreadQueue for NS \"%s\"", ns->idstr);
-      repackstreamer_abort(rman->gstate.rpst);
-      rman->gstate.rpst = NULL;
-      resourcelog_term(&rman->gstate.rlog, NULL, 1);
-      resourceinput_purge(&rman->gstate.rinput);
-      config_abandonposition(&rman->gstate.pos);
-      return -1;
-   }
-   // ensure all threads have initialized
-   if (tq_check_init(rman->tq)) {
-      LOG(LOG_ERR, "Some threads failed to initialize for NS \"%s\"\n", ns->idstr);
-      snprintf(response->errorstr, MAX_ERROR_BUFFER, "Some threads failed to initialize for NS \"%s\"", ns->idstr);
-      tq_set_flags(rman->tq, TQ_ABORT);
-      repackstreamer_abort(rman->gstate.rpst);
-      rman->gstate.rpst = NULL;
-      resourcelog_term(&rman->gstate.rlog, NULL, 1);
-      resourceinput_purge(&rman->gstate.rinput);
-      config_abandonposition(&rman->gstate.pos);
-      return -1;
-   }
-   LOG(LOG_INFO, "Rank %zu is now targetting NS \"%s\"\n", rman->ranknum, ns->idstr);
-   return 0;
-}
-
-/**
- * Calculate the min and max values for the given ref distribution of the NS
- * @param marfs_ns* ns : Namespace to split ref ranges across
- * @param size_t workingranks : Total number of operating ranks
- * @param size_t refdist : Reference distribution index
- * @param size_t* refmin : Reference to be populated with the minimum range value
- * @param size_t* refmin : Reference to be populated with the maximum range value (non-inclusive)
- */
-static void getNSrange(marfs_ns* ns, size_t workingranks, size_t refdist, size_t* refmin, size_t* refmax) {
-   size_t refperrank = ns->prepo->metascheme.refnodecount / workingranks; // 'average' number of reference ranges per rank (truncated to integer)
-   size_t remainder = ns->prepo->metascheme.refnodecount % workingranks;  // 'remainder' ref dirs, omitted if every rank merely got the 'average'
-   size_t prevextra = (refdist > remainder) ? remainder : refdist; // how many 'remainder' dirs were included in previous distributions
-   *refmin = (refdist * refperrank) + prevextra; // include the 'average' per-rank count, plus any 'remainder' dirs already issued
-   *refmax = (*refmin + refperrank) + ((refdist < remainder) ? 1 : 0); // add on the 'average' and one extra, if 'remainder' dirs still exist
-   LOG(LOG_INFO, "Using Min=%zu / Max=%zu for ref distribution %zu on NS \"%s\"\n", *refmin, *refmax, refdist, ns->idstr);
-   return;
-}
-
-/**
- * Populate the current rank's oldlogs HASH_TABLE
- * @param rmanstate* rman : State of this rank
- * @param const char* scanroot : Logroot, below which to scan
- * @return int : Zero on success, or -1 on failure
- */
-static int findoldlogs(rmanstate* rman, const char* scanroot, time_t skipthresh) {
-   // construct a HASH_TABLE to hold old logfile references
-   HASH_NODE* lognodelist = calloc(rman->nscount, sizeof(struct hash_node_struct));
-   if (lognodelist == NULL) {
-      LOG(LOG_ERR, "Failed to allocate a log node list of length %zu\n", rman->nscount);
-      return -1;
-   }
-   size_t pindex = 0;
-   while (pindex < rman->nscount) {
-      (lognodelist + pindex)->name = strdup((rman->nslist[pindex])->idstr);
-      if ((lognodelist + pindex)->name == NULL) {
-         LOG(LOG_ERR, "Failed to duplicate NS ID String: \"%s\"\n", (rman->nslist[pindex])->idstr);
-         while (pindex > 0) { pindex--; free((lognodelist + pindex)->name); free((lognodelist + pindex)->content); }
-         free(lognodelist);
-         return -1;
-      }
-      // node weight should be pre-zeroed by calloc()
-      (lognodelist + pindex)->content = calloc(1, sizeof(loginfo));
-      if ((lognodelist + pindex)->content == NULL) {
-         LOG(LOG_ERR, "Failed to allocate a loginfo struct (%s)\n", strerror(errno));
-         free((lognodelist + pindex)->name);
-         while (pindex > 0) { pindex--; free((lognodelist + pindex)->name); free((lognodelist + pindex)->content); }
-         free(lognodelist);
-         return -1;
-      }
-      loginfo* newinfo = (loginfo*)((lognodelist + pindex)->content);
-      newinfo->nsindex = pindex;
-      // logcount and requests are pre-zeroed by calloc()
-      pindex++;
-   }
-   if ((rman->oldlogs = hash_init(lognodelist, rman->nscount, 1)) == NULL) {
-      LOG(LOG_ERR, "Failed to initialize old logfile hash table\n");
-      while (pindex > 0) { pindex--; free((lognodelist + pindex)->name); free((lognodelist + pindex)->content); }
-      free(lognodelist);
-      return -1;
-   }
-   // NOTE -- Once the table is initialized, this func won't bother to free it on error.
-   //         That is now the caller's responsibility.
-   DIR*  dirlist[3] = {0};
-   workrequest request = {
-      .type = RLOG_WORK,
-      .nsindex = 0,
-      .refdist = 0,
-      .iteration = {0},
-      .ranknum = 0
-   };
-   // open the 'scanroot'
-   dirlist[0] = opendir(scanroot);
-   if (dirlist[0] == NULL) {
-      LOG(LOG_ERR, "Failed to open logging root \"%s\" for scanning\n", scanroot);
-      return -1;
-   }
-   size_t logcount = 0; // per iteration log count
-   size_t logdepth = 1;
-   char itercheck = 0;
-   loginfo* linfo = NULL; // for caching a target location for new log request info
-   while (logdepth) {
-      if (logdepth == 1) { // identify the iteration to tgt
-         if (rman->execprevroot) {
-            // check for potential exit
-            if (itercheck) {
-               LOG(LOG_INFO, "Preparing to exit, after completing scan of prevexec iteration: \"%s/%s\"\n", scanroot, rman->iteration);
-               request.iteration[0] = '\0';
-               logdepth--;
-               continue;
-            }
-            // note a pass over this section, so we never repeat
-            itercheck = 1;
-            // only targeting a matching iteration
-            if (snprintf(request.iteration, ITERATION_STRING_LEN, "%s", rman->iteration) < 1) {
-               LOG(LOG_ERR, "Failed to populate request with running iteration \"%s\"\n", rman->iteration);
-               return -1;
-            }
-         }
-         else {
-            // scan through the root dir, looking for other iterations
-            int olderr = errno;
-            errno = 0;
-            struct dirent* entry;
-            while ((entry = readdir(dirlist[0]))) {
-               // ignore '.'-prefixed entries
-               if (strncmp(entry->d_name, ".", 1) == 0) { continue; }
-               // ignore the running iteration name, as that will cause all kinds of problems
-               if (strncmp(entry->d_name, rman->iteration, ITERATION_STRING_LEN) == 0) {
-                  LOG(LOG_INFO, "Skipping active iteration path: \"%s\"\n", rman->iteration);
-                  continue;
-               }
-               // all other entries are assumed to be valid iteration targets
-               if (snprintf(request.iteration, ITERATION_STRING_LEN, "%s", entry->d_name) >= ITERATION_STRING_LEN) {
-                  LOG(LOG_ERR, "Failed to populate request string for old iteration: \"%s\"\n", entry->d_name);
-                  return -1;
-               }
-               break;
-            }
-            if (errno) {
-               LOG(LOG_ERR, "Failed to read log root dir: \"%s\" (%s)\n", scanroot);
-               return -1;
-            }
-            errno = olderr;
-            if (entry == NULL) {
-               // no iterations remain, we are done
-               closedir(dirlist[0]);
-               dirlist[0] = NULL; // just to be explicit about clearing this value
-               request.iteration[0] = '\0';
-               logdepth--;
-               LOG(LOG_INFO, "Preparing to exit, after completing scan of previous iterations under \"%s\"\n", scanroot);
-               continue;
-            }
-         }
-         // open the dir of the matching iteration
-         int newfd = openat(dirfd(dirlist[0]), request.iteration, O_DIRECTORY | O_RDONLY, 0);
-         if (newfd < 0) {
-            LOG(LOG_ERR, "Failed to open log root for iteration: \"%s\" (%s)\n", request.iteration, strerror(errno));
-            return -1;
-         }
-         // check if program args are compatible with this run
-         if (!rman->execprevroot) {
-            struct stat stval; // for checking the age of previous iteration files
-            int sumfd = openat(newfd, SUMMARY_FILENAME, O_RDONLY, 0);
-            if (sumfd < 0) {
-               LOG(LOG_WARNING, "Failed to open summary log for old iteration: \"%s\" (%s)\n",
-                    request.iteration, strerror(errno));
-               // stat the root of this iteration
-               if (fstat(newfd, &stval)) {
-                  LOG(LOG_ERR, "Failed to stat root of old iteration: \"%s\" (%s)\n",
-                                request.iteration, strerror(errno));
-                  close(newfd);
-                  return -1;
-               }
-               // check if this iteration is old enough for us to feel comfortable running
-               if (stval.st_mtime > skipthresh) {
-                  fprintf(stderr, "ERROR: Found recent iteration with possible conflicting operations (no summary logfile): \"%s\"\n",
-                                   request.iteration);
-                  close(newfd);
-                  return -1;
-               }
-               // otherwise, assume we can safely skip over this iteration dir
-               printf("Skipping over previous iteration lacking a summary logfile: \"%s\"\n", request.iteration);
-               continue;
-            }
-            // stat the summary logfile
-            if (fstat(sumfd, &stval)) {
-               LOG(LOG_ERR, "Failed to stat summary logfile of old iteration: \"%s\" (%s)\n",
-                             request.iteration, strerror(errno));
-               close(sumfd);
-               close(newfd);
-               return -1;
-            }
-            // for runs actually modifying FS content (non-quota-only runs),
-            //    check if this iteration is old enough for us to feel comfortable running
-            if ((rman->gstate.thresh.gcthreshold  ||  rman->gstate.thresh.repackthreshold  ||
-                   rman->gstate.thresh.rebuildthreshold  ||  rman->gstate.thresh.cleanupthreshold)
-                  &&  stval.st_mtime > rman->gstate.thresh.cleanupthreshold) {
-               fprintf(stderr, "ERROR: Found recent iteration with possible conflicting operations: \"%s\"\n", request.iteration);
-               close(newfd);
-               return -1;
-            }
-            // open a file stream for the summary logfile
-            FILE* summaryfile = fdopen(sumfd, "r");
-            if (summaryfile == NULL) {
-               LOG(LOG_ERR, "Failed to open a file stream for summary log for old iteration: \"%s\" (%s)\n",
-                    request.iteration, strerror(errno));
-               close(sumfd);
-               close(newfd);
-               return -1;
-            }
-            // parse args into a temporary state struct, for comparison
-            rmanstate tmpstate;
-            memset(&tmpstate, 0, sizeof(tmpstate));
-            tmpstate.config = rman->config;
-            int parseargres = parse_program_args(&tmpstate, summaryfile);
-            fclose(summaryfile);
-            if (parseargres > 0) {
-               printf("Skipping over previous iteration with incompatible config: \"%s\"\n", request.iteration);
-               continue;
-            }
-            if (parseargres) {
-               LOG(LOG_ERR, "Failed to parse arguments from summary log of old iteration: \"%s\" (%s)\n",
-                    request.iteration, strerror(errno));
-               close(newfd);
-               return -1;
-            }
-            // check if previous run args are a superset if this run's
-            if (tmpstate.gstate.thresh.gcthreshold > rman->gstate.thresh.gcthreshold  ||
-                 tmpstate.gstate.thresh.repackthreshold > rman->gstate.thresh.repackthreshold  ||
-                 tmpstate.gstate.thresh.rebuildthreshold > rman->gstate.thresh.rebuildthreshold  ||
-                 tmpstate.gstate.thresh.cleanupthreshold > rman->gstate.thresh.cleanupthreshold  ||
-                 tmpstate.gstate.dryrun != rman->gstate.dryrun) {
-               printf("Skipping over previous iteration with incompatible arguments: \"%s\"\n", request.iteration);
-               continue;
-            }
-         }
-         dirlist[1] = fdopendir(newfd);
-         if (dirlist[1] == NULL) {
-            LOG(LOG_ERR, "Failed to open DIR reference for previous iteration log root: \"%s\" (%s)\n",
-                 request.iteration, strerror(errno));
-            close(newfd);
-            return -1;
-         }
-         logdepth++;
-         LOG(LOG_INFO, "Entered iteration dir: \"%s\"\n", request.iteration);
-         logcount = 0;
-      }
-      else if (logdepth == 2) { // identify the NS subdirs
-         // scan through the iteration dir, looking for namespaces
-         int olderr = errno;
-         errno = 0;
-         struct dirent* entry;
-         while ((entry = readdir(dirlist[1]))) {
-            // ignore '.'-prefixed entries
-            if (strncmp(entry->d_name, ".", 1) == 0) { continue; }
-            // ignore the summary log
-            if (strcmp(entry->d_name, SUMMARY_FILENAME) == 0) { continue; }
-            // all other entries are assumed to be namespaces
-            break;
-         }
-         if (errno) {
-            LOG(LOG_ERR, "Failed to read iteration dir: \"%s\" (%s)\n", request.iteration, strerror(errno));
-            return -1;
-         }
-         errno = olderr;
-         if (entry == NULL) {
-            // no Namespaces remain, we are done
-            if (logcount == 0  &&  rman->execprevroot == NULL) {
-               // if this is a dead iteration, with no logs remaining, just delete it
-               if (unlinkat(dirfd(dirlist[1]), SUMMARY_FILENAME, 0)) {
-                  fprintf(stderr, "WARNING: Failed to remove summary log of previous iteration: \"%s\"\n", request.iteration);
-               }
-               else if (unlinkat(dirfd(dirlist[0]), request.iteration, AT_REMOVEDIR)) {
-                  fprintf(stderr, "WARNING: Failed to remove previous iteration log root: \"%s\"\n", request.iteration);
-               }
-            }
-            closedir(dirlist[1]);
-            dirlist[1] = NULL; // just to be explicit about clearing this value
-            request.nsindex = 0;
-            logdepth--;
-            LOG(LOG_INFO, "Finished processing iteration dir: \"%s\"\n", request.iteration);
-            if (logcount) {
-               printf("This run will incorporate %zu logfiles from: \"%s/%s\"\n", logcount, scanroot, request.iteration);
-            }
-            continue;
-         }
-         // check what index this NS corresponds to
-         char* oldnsid = strdup(entry->d_name);
-         if (oldnsid == NULL) {
-            LOG(LOG_ERR, "Failed to duplicate dirent name: \"%s\"\n", entry->d_name);
-            return -1;
-         }
-         char* tmpparse = oldnsid;
-         while (*tmpparse != '\0') { if (*tmpparse == '#') { *tmpparse = '/'; } tmpparse++; }
-         HASH_NODE* lnode = NULL;
-         if (hash_lookup(rman->oldlogs, oldnsid, &lnode)) {
-            // if we can't map this to a NS entry, just ignore it
-            fprintf(stderr, "WARNING: Encountered resource logs from a previoius iteration (\"%s\")\n"
-                             "         associated with an unrecognized namespace (\"%s\").\n"
-                             "         These logfiles will be ignored by the current run.\n"
-                             "         See \"%s/%s/%s\"\n", request.iteration, entry->d_name, scanroot, request.iteration, entry->d_name);
-            free(oldnsid);
-         }
-         else {
-            free(oldnsid);
-            // pull the NS index value from the corresponding hash node
-            linfo = (loginfo*)(lnode->content);
-            request.nsindex = linfo->nsindex;
-            // open the corresponding subdir and traverse into it
-            int newfd = openat(dirfd(dirlist[1]), entry->d_name, O_DIRECTORY | O_RDONLY, 0);
-            if (newfd < 0) {
-               LOG(LOG_ERR, "Failed to open log NS subdir: \"%s/%s/%s\" (%s)\n", scanroot, request.iteration, entry->d_name, strerror(errno));
-               return -1;
-            }
-            dirlist[2] = fdopendir(newfd);
-            if (dirlist[2] == NULL) {
-               LOG(LOG_ERR, "Failed to open DIR reference for previous iteration log root: \"%s\" (%s)\n",
-                    request.iteration, strerror(errno));
-               close(newfd);
-               return -1;
-            }
-            logdepth++;
-            continue;
-         }
-      }
-      else { // identify the rank to tgt
-         // scan through the NS dir, looking for individual logfiles
-         int olderr = errno;
-         errno = 0;
-         struct dirent* entry;
-         while ((entry = readdir(dirlist[2]))) {
-            // ignore '.'-prefixed entries
-            if (strncmp(entry->d_name, ".", 1) == 0) { continue; }
-            // ignore error-prefixed entries
-            if (strncmp(entry->d_name, ERROR_LOG_PREFIX, strlen(ERROR_LOG_PREFIX)) == 0) { continue; }
-            // all other entries are assumed to be logfiles
-            break;
-         }
-         if (errno) {
-            LOG(LOG_ERR, "Failed to read NS dir: \"%s/%s/%s\" (%s)\n", scanroot, request.iteration,
-                 (rman->nslist[request.nsindex])->idstr, strerror(errno));
-            return -1;
-         }
-         errno = olderr;
-         if (entry == NULL) {
-            // no iterations remain, we are done
-            closedir(dirlist[2]);
-            dirlist[2] = NULL; // just to be explicit about clearing this value
-            request.ranknum = 0;
-            logdepth--;
-            continue;
-         }
-         // identify the rank number associated with this logfile
-         char* fparse = entry->d_name;
-         while (*fparse != '\0') {
-            if (*fparse == '-') { fparse++; break; }
-            fparse++;
-         }
-         char* endptr = NULL;
-         unsigned long long parseval = strtoull(fparse, &endptr, 10);
-         if (endptr == NULL  ||  *endptr != '\0'  ||  *fparse == '\0'  ||  parseval >= SIZE_MAX  ||  parseval >= ULLONG_MAX) {
-            fprintf(stderr, "WARNING: Failed to identify rank number associated with logfile \"%s/%s/%s/%s\"\n"
-                             "         The logfile will be skipped\n",
-                             scanroot, request.iteration, (rman->nslist[request.nsindex])->idstr, entry->d_name);
-            continue;
-         }
-         request.ranknum = (size_t)parseval;
-         // populate the HASH_TABLE entry with the newly completed request
-         if (linfo->logcount % OLDLOG_PREALLOC == 0) {
-            // logs are allocated in groups of OLDLOG_PREALLOC, so we can use modulo to check for overflow
-            workrequest* newrequests = realloc(linfo->requests, (linfo->logcount + OLDLOG_PREALLOC) * sizeof(*newrequests));
-            if (newrequests == NULL) {
-               LOG(LOG_ERR, "Failed to allocate a list of old log requests with length %zu\n", linfo->logcount + OLDLOG_PREALLOC);
-               return -1;
-            }
-            linfo->requests = newrequests;
-         }
-         LOG(LOG_INFO, "Detected old logfile for iteration \"%s\", NS \"%s\", Rank %zu\n",
-              request.iteration, (rman->nslist[request.nsindex])->idstr, request.ranknum);
-         *(linfo->requests + linfo->logcount) = request; // NOTE -- I believe this assignment is safe,
-                                                         //         because .iteration is a static char array
-         linfo->logcount++;
-         logcount++;
-      }
-   }
-   LOG(LOG_INFO, "Finished scan of old resource logs below \"%s\"\n", scanroot);
-   return 0;
-}
-
-/**
- * Process the given request
- * @param rmanstate* rman : Current resource manager rank state
- * @param workrequest* request : Request to process
- * @param workresponse* response : Response to populate
- * @return int : Zero if the rank should send the populated response and terminate
- *               One if the rank should send the populated response and continue processing
- *               -1 if the rank should send the populated response, then abort
- */
-static int handlerequest(rmanstate* rman, workrequest* request, workresponse* response) {
-   // pre-populate response with a 'fatal error' condition, just in case
-   response->request = *request;
-   response->haveinfo = 0;
-   memset(&response->report, 0, sizeof(response->report));
-   memset(&response->summary, 0, sizeof(response->summary));
-   response->errorlog = 0;
-   response->fatalerror = 1;
-   snprintf(response->errorstr, MAX_ERROR_BUFFER, "UNKNOWN-ERROR!");
-   // identify and process the request
-   if (request->type == RLOG_WORK) {
-      // potentially update our state to target the NS
-      if (rman->gstate.rlog == NULL) {
-         if (setranktgt(rman, rman->nslist[request->nsindex], response)) {
-            LOG(LOG_ERR, "Failed to update target of rank %zu to NS \"%s\"\n",
-                          rman->ranknum, rman->nslist[request->nsindex]->idstr);
-            // leave the errorstr alone, as the helper func will have left a more descriptive message
-            return -1;
-         }
-      }
-      // identify the path of the rlog
-      const char* tmplogroot = rman->logroot;
-      if (rman->execprevroot) { tmplogroot = rman->execprevroot; }
-      char* rlogpath = resourcelog_genlogpath(0, tmplogroot, request->iteration,
-                                                  rman->nslist[request->nsindex], request->ranknum);
-      if (rlogpath == NULL) {
-         LOG(LOG_ERR, "Failed to generate logpath for NS \"%s\" ranknum \"%s\"\n",
-                       rman->nslist[request->nsindex]->idstr, request->ranknum);
-         snprintf(response->errorstr, MAX_ERROR_BUFFER, "Failed to generate logpath for NS \"%s\" ranknum \"%zu\"",
-                   rman->nslist[request->nsindex]->idstr, request->ranknum);
-         return -1;
-      }
-      // determine what we're actually doing with this logfile
-      if (rman->execprevroot) {
-         // we are using the rlog as an input
-         if (resourceinput_setlogpath(&rman->gstate.rinput, rlogpath)) {
-            LOG(LOG_ERR, "Failed to update rank %zu input to logfile \"%s\"\n", rman->ranknum, rlogpath);
-            snprintf(response->errorstr, MAX_ERROR_BUFFER, "Failed to update rank %zu input to logfile \"%s\"",
-                      rman->ranknum, rlogpath);
-            return -1;
-         }
-         // wait for our input to be exhausted (don't respond until we are ready for more work)
-         if (resourceinput_waitforcomp(&rman->gstate.rinput)) {
-            LOG(LOG_ERR, "Failed to wait for completion of logfile \"%s\"\n", request->refdist);
-            snprintf(response->errorstr, MAX_ERROR_BUFFER, "Failed to wait for completion of logfile \"%s\"", rlogpath);
-            return -1;
-         }
-      }
-      else {
-         // we are just cleaning up this old logfile by replaying it into our current one
-         // open the specified rlog for read
-         RESOURCELOG newlog = NULL;
-         if (resourcelog_init(&newlog, rlogpath, RESOURCE_READ_LOG, rman->nslist[request->nsindex])) {
-            const int err = errno;
-            LOG(LOG_ERR, "Failed to open logfile for read: \"%s\" (%s)\n", rlogpath, strerror(err));
-            snprintf(response->errorstr, MAX_ERROR_BUFFER, "Failed to open logfile for read: \"%s\" (%s)",
-                      rlogpath, strerror(err));
-            free(rlogpath);
-            return -1;
-         }
-         if (resourcelog_replay(&newlog, &rman->gstate.rlog, NULL)) {
-            LOG(LOG_ERR, "Failed to replay logfile \"%s\" into active state\n", rlogpath);
-            snprintf(response->errorstr, MAX_ERROR_BUFFER, "Failed to replay logfile \"%s\"", rlogpath);
-            resourcelog_abort(&newlog);
-            free(rlogpath);
-            return -1;
-         }
-      }
-      free(rlogpath);
-   }
-   else if (request->type == NS_WORK) {
-      // potentially update our state to target the NS
-      if (rman->gstate.rlog == NULL) {
-         if (setranktgt(rman, rman->nslist[request->nsindex], response)) {
-            LOG(LOG_ERR, "Failed to update target of rank %zu to NS \"%s\"\n",
-                          rman->ranknum, rman->nslist[request->nsindex]->idstr);
-            // leave the errorstr alone, as the helper func will have left a more descriptive message
-            return -1;
-         }
-      }
-      // calculate the start and end reference indices for this request
-      size_t refmin = -1;
-      size_t refmax = -1;
-      getNSrange(rman->nslist[request->nsindex], rman->workingranks, request->refdist, &refmin, &refmax);
-      // only actually perform the work if it is a valid reference range
-      // NOTE -- This is to handle a case where the number of NS ref dirs < the number of resource manager worker ranks
-      if (refmax != refmin) {
-         // update our input to reference the new target range
-         if (resourceinput_setrange(&rman->gstate.rinput, refmin, refmax)) {
-            LOG(LOG_ERR, "Failed to set NS \"%s\" reference range values for distribution %zu\n",
-                 rman->nslist[request->nsindex]->idstr, request->refdist);
-            snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                      "Failed to set NS \"%s\" reference range values for distribution %zu",
-                      rman->nslist[request->nsindex]->idstr, request->refdist);
-            return -1;
-         }
-         // wait for our input to be exhausted (don't respond until we are ready for more work)
-         if (resourceinput_waitforcomp(&rman->gstate.rinput)) {
-            LOG(LOG_ERR, "Failed to wait for completion of NS \"%s\" reference distribution %zu\n",
-                 rman->nslist[request->nsindex]->idstr, request->refdist);
-            snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                      "Failed to wait for completion of NS \"%s\" reference distribution %zu",
-                      rman->nslist[request->nsindex]->idstr, request->refdist);
-            return -1;
-         }
-      }
-      else {
-         // NOTE -- In this case, refmin and refmax should both equal the total refdir count
-         LOG(LOG_INFO, "Skipping distribution %zu of NS \"%s\" (%zu of %zu dirs already processed)\n",
-              request->refdist, rman->nslist[request->nsindex]->idstr, refmin, refmax);
-      }
-   }
-   else if (request->type == COMPLETE_WORK) {
-      // complete any outstanding work within the current NS
-      if (rman->gstate.rlog == NULL) {
-         LOG(LOG_ERR, "Rank %zu was asked to complete work, but has none\n", rman->ranknum);
-         snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                   "Rank %zu was asked to complete work, but has none", rman->ranknum);
-         return -1;
-      }
-      // terminate the resource input
-      if (rman->gstate.rinput  &&  resourceinput_term(&rman->gstate.rinput)) {
-         LOG(LOG_ERR, "Failed to terminate resourceinput\n");
-         snprintf(response->errorstr, MAX_ERROR_BUFFER, "Failed to terminate resourceinput");
-         // don't quit out yet, try to collect thread states first
-      }
-      char threaderror = 0;
-      if (rman->tq) {
-         // wait for the queue to be marked as FINISHED
-         TQ_Control_Flags setflags = 0;
-         if (tq_wait_for_flags(rman->tq, 0, &setflags)) {
-            LOG(LOG_ERR, "Failed to wait on ThreadQueue state flags\n");
-            snprintf(response->errorstr, MAX_ERROR_BUFFER, "Failed to wait on ThreadQueue state flags");
-            return -1;
-         }
-         if (setflags != TQ_FINISHED) {
-            LOG(LOG_WARNING, "Unexpected (NON-FINISHED) ThreadQueue state flags: %u\n", setflags);
-            threaderror = 1;
-         }
-         else {
-            // wait for TQ completion
-            if (tq_wait_for_completion(rman->tq)) {
-               LOG(LOG_ERR, "Failed to wait for ThreadQueue completion\n");
-               threaderror = 1;
-            }
-         }
-         // gather all thread status values
-         rthread_state* tstate = NULL;
-         int retval;
-         while ((retval = tq_next_thread_status(rman->tq, (void**)&tstate)) > 0) {
-            LOG(LOG_INFO, "Got state for Thread %zu\n", tstate->tID);
-            // verify thread status
-            if (tstate == NULL) {
-               LOG(LOG_ERR, "Rank %zu encountered NULL thread state when completing NS \"%s\"\n",
-                    rman->ranknum, rman->gstate.pos.ns->idstr);
-               snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                         "Rank %zu encountered NULL thread state when completing NS \"%s\"",
-                         rman->ranknum, rman->gstate.pos.ns->idstr);
-               free(tstate);
-               return -1;
-            }
-            if (tstate->fatalerror) {
-               LOG(LOG_ERR, "Fatal Error in NS \"%s\": \"%s\"\n",
-                    rman->gstate.pos.ns->idstr, tstate->errorstr);
-               snprintf(response->errorstr, MAX_ERROR_BUFFER, "Fatal Error in NS \"%s\": \"%s\"",
-                         rman->gstate.pos.ns->idstr, tstate->errorstr);
-               free(tstate);
-               return -1;
-            }
-            response->report.fileusage   += tstate->report.fileusage;
-            response->report.byteusage   += tstate->report.byteusage;
-            response->report.filecount   += tstate->report.filecount;
-            response->report.objcount    += tstate->report.objcount;
-            response->report.bytecount   += tstate->report.bytecount;
-            response->report.streamcount += tstate->report.streamcount;
-            response->report.delobjs     += tstate->report.delobjs;
-            response->report.delfiles    += tstate->report.delfiles;
-            response->report.delstreams  += tstate->report.delstreams;
-            response->report.volfiles    += tstate->report.volfiles;
-            response->report.rpckfiles   += tstate->report.rpckfiles;
-            response->report.rpckbytes   += tstate->report.rpckbytes;
-            response->report.rbldobjs    += tstate->report.rbldobjs;
-            response->report.rbldbytes   += tstate->report.rbldbytes;
-            free(tstate);
-         }
-         if (retval) {
-            LOG(LOG_ERR, "Failed to collect thread states during completion of work on NS \"%s\"\n",
-                 rman->gstate.pos.ns->idstr);
-            snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                      "Failed to collect thread status during completion of work on NS \"%s\"",
-                      rman->gstate.pos.ns->idstr);
-            return -1;
-         }
-         response->haveinfo = 1; // note that we have info for the manager to process
-         // close our the thread queue, repack streamer, and logs
-         if (tq_close(rman->tq)) {
-            LOG(LOG_ERR, "Failed to close ThreadQueue after completion of work on NS \"%s\"\n",
-                 rman->gstate.pos.ns->idstr);
-            snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                      "Failed to close ThreadQueue after completion of work on NS \"%s\"",
-                      rman->gstate.pos.ns->idstr);
-            return -1;
-         }
-         rman->tq = NULL;
-      }
-      // destroy the resource input, if present
-      if (rman->gstate.rinput) {
-         if (resourceinput_destroy(&rman->gstate.rinput)) {
-            // nothing to do but complain
-            LOG(LOG_WARNING, "Failed to destroy resourceinput structure\n");
-            rman->gstate.rinput = NULL; // NULL it out anyhow
-         }
-      }
-      // need to preserve our logfile, allowing the manager to remove it
-      char* outlogpath = resourcelog_genlogpath(0, rman->logroot, rman->iteration, rman->gstate.pos.ns, rman->ranknum);
-      int rlogret;
-      if ((rlogret = resourcelog_term(&rman->gstate.rlog, &response->summary, 0)) < 0) {
-         LOG(LOG_ERR, "Failed to terminate log \"%s\" following completion of work on NS \"%s\"\n",
-              outlogpath, rman->gstate.pos.ns->idstr);
-         snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                   "Failed to terminate log \"%s\" following completion of work on NS \"%s\"",
-                   outlogpath, rman->gstate.pos.ns->idstr);
-         free(outlogpath);
-         return -1;
-      }
-      if (outlogpath) { free(outlogpath); }
-      if (rlogret) { response->errorlog = 1; } // note if our log was preserved due to errors being present
-      if (rman->gstate.rpst) {
-         if (repackstreamer_complete(rman->gstate.rpst)) {
-            LOG(LOG_ERR, "Failed to complete repack streamer during completion of NS \"%s\"\n", rman->gstate.pos.ns->idstr);
-            snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                      "Failed to complete repack streamer during completion of NS \"%s\"", rman->gstate.pos.ns->idstr);
-            return -1;
-         }
-         rman->gstate.rpst = NULL;
-      }
-      const char* nsidstr = rman->gstate.pos.ns->idstr;
-      if (config_abandonposition(&rman->gstate.pos)) {
-         LOG(LOG_ERR, "Failed to abandon marfs position during completion of NS \"%s\"\n", nsidstr);
-         snprintf(response->errorstr, MAX_ERROR_BUFFER,
-                   "Failed to abandon marfs position during completion of NS \"%s\"", nsidstr);
-         return -1;
-      }
-      if (threaderror) {
-         snprintf(response->errorstr, MAX_ERROR_BUFFER, "ThreadQueue had unexpected termination condition");
-         return -1;
-      }
-   }
-   else if (request->type == TERMINATE_WORK) {
-      // simply note and terminate
-      response->fatalerror = 0;
-      return 0;
-   }
-   else { // assume ABORT_WORK
-      // clear our fatal error state, but indicate to abort after sending the response
-      response->fatalerror = 0;
-      return -1;
-   }
-   // clear our fatal error state
-   response->fatalerror = 0;
-   // indicate that we have a response to send
-   return 1;
-}
-
-/**
- * Process the given response
- * @param rmanstate* rman : Rank state
- * @param size_t ranknum : Responding rank number
- * @param workresponse* response : Response to process
- * @param workrequest* request : New request to populate
- * @return int : A positive value if a new request has been populated;
- *               0 if no request should be sent (rank has exited, or hit a fatal error);
- *               -1 if a fatal error has occurred in this function itself (invalid processing)
- */
-static int handleresponse(rmanstate* rman, size_t ranknum, workresponse* response, workrequest* request) {
-   // check for a fatal error condition, as this overrides all other behaviors
-   if (response->fatalerror) {
-      // note a fatalerror, and don't generate a request
-      fprintf(stderr, "ERROR: Rank %zu%s hit a fatal error condition\n"
-                       "       All ranks will now attempt to terminate\n"
-                       "       Work Type --\n"
-                       "         %s\n"
-                       "       Error Description --\n"
-                       "         \"%s\"\n",
-                       ranknum, (ranknum == rman->ranknum) ? " (this rank)" : "",
-                       (response->request.type == RLOG_WORK)      ? "Resource Log Processing" :
-                       (response->request.type == NS_WORK)        ? "Namespace Processing"    :
-                       (response->request.type == COMPLETE_WORK)  ? "Namespace Finalization"  :
-                       (response->request.type == TERMINATE_WORK) ? "Termination"             :
-                       (response->request.type == ABORT_WORK)     ? "Abort Condition"         :
-                       "UNKNOWN WORK TYPE",
-                       response->errorstr);
-      rman->fatalerror = 1; // set our fatal error condition, so we signal future workers to terminate
-      rman->terminatedworkers[ranknum] = 1;
-      return 0;
-   }
-   if (rman->fatalerror  &&  (response->request.type != ABORT_WORK)) {
-      // if we are in an error state, all ranks should be signaled to abort
-      LOG(LOG_INFO, "Signaling Rank %zu to Abort\n", ranknum);
-      request->type = ABORT_WORK;
-      request->nsindex = 0;
-      request->refdist = 0;
-      request->iteration[0] = '\0';
-      request->ranknum = 0;
-      return 1;
-   }
-   // handle the response based on type
-   if (response->request.type == TERMINATE_WORK  ||  response->request.type == ABORT_WORK) {
-      // these simply indicate that the rank has terminated
-      rman->terminatedworkers[ranknum] = 1;
-      printf("  Rank %zu is terminating\n", ranknum);
-      return 0;
-   }
-   else if (response->request.type == COMPLETE_WORK) {
-      if (response->request.nsindex != rman->nscount) { // only perform processing / cleanup for real completions
-         printf("  Rank %zu completed work on NS \"%s\"\n", ranknum, rman->nslist[response->request.nsindex]->idstr);
-         // possibly process info from the rank
-         if (response->haveinfo) {
-            // incorporate walk report
-            rman->walkreport[response->request.nsindex].fileusage   += response->report.fileusage;
-            rman->walkreport[response->request.nsindex].byteusage   += response->report.byteusage;
-            rman->walkreport[response->request.nsindex].filecount   += response->report.filecount;
-            rman->walkreport[response->request.nsindex].objcount    += response->report.objcount;
-            rman->walkreport[response->request.nsindex].bytecount   += response->report.bytecount;
-            rman->walkreport[response->request.nsindex].streamcount += response->report.streamcount;
-            rman->walkreport[response->request.nsindex].delobjs     += response->report.delobjs;
-            rman->walkreport[response->request.nsindex].delfiles    += response->report.delfiles;
-            rman->walkreport[response->request.nsindex].delstreams  += response->report.delstreams;
-            rman->walkreport[response->request.nsindex].volfiles    += response->report.volfiles;
-            rman->walkreport[response->request.nsindex].rpckfiles   += response->report.rpckfiles;
-            rman->walkreport[response->request.nsindex].rpckbytes   += response->report.rpckbytes;
-            rman->walkreport[response->request.nsindex].rbldobjs    += response->report.rbldobjs;
-            rman->walkreport[response->request.nsindex].rbldbytes   += response->report.rbldbytes;
-            // incorporate log summary
-            rman->logsummary[response->request.nsindex].deletion_object_count += response->summary.deletion_object_count;
-            rman->logsummary[response->request.nsindex].deletion_object_failures += response->summary.deletion_object_failures;
-            rman->logsummary[response->request.nsindex].deletion_reference_count += response->summary.deletion_reference_count;
-            rman->logsummary[response->request.nsindex].deletion_reference_failures += response->summary.deletion_reference_failures;
-            rman->logsummary[response->request.nsindex].rebuild_count += response->summary.rebuild_count;
-            rman->logsummary[response->request.nsindex].rebuild_failures += response->summary.rebuild_failures;
-            rman->logsummary[response->request.nsindex].repack_count += response->summary.repack_count;
-            rman->logsummary[response->request.nsindex].repack_failures += response->summary.repack_failures;
-            // output all gathered info, prior to possible log deletion
-            outputinfo(rman->summarylog, rman->nslist[response->request.nsindex], &response->report , &response->summary);
-         }
-         // identify the output log of the transmitting rank
-         char* outlogpath = resourcelog_genlogpath(0, rman->logroot, rman->iteration,
-                                                    rman->nslist[response->request.nsindex], ranknum);
-         if (outlogpath == NULL) {
-            fprintf(stderr, "ERROR: Failed to identify the output log path of Rank %zu for NS \"%s\"\n",
-                     ranknum, rman->nslist[response->request.nsindex]->idstr);
-            rman->fatalerror = 1;
-            return -1;
-         }
-         // potentially duplicate the logfile to a final location
-         if (rman->preservelogtgt) {
-            char* preslogpath = resourcelog_genlogpath(1, rman->preservelogtgt, rman->iteration,
-                                                        rman->nslist[response->request.nsindex], ranknum);
-            if (preslogpath == NULL) {
-               fprintf(stderr, "ERROR: Failed to identify the preserve log path of Rank %zu for NS \"%s\"\n",
-                        ranknum, rman->nslist[response->request.nsindex]->idstr);
-               free(outlogpath);
-               rman->fatalerror = 1;
-               return -1;
-            }
-            // simply use a hardlink for this purpose, no need to make a 'real' duplicate
-            if (link(outlogpath, preslogpath)) {
-               fprintf(stderr, "ERROR: Failed to link logfile \"%s\" to new target path: \"%s\"\n",
-                        outlogpath, preslogpath);
-               free(preslogpath);
-               free(outlogpath);
-               rman->fatalerror = 1;
-               return -1;
-            }
-            free(preslogpath);
-         }
-         // open the logfile for read
-         RESOURCELOG ranklog = NULL;
-         if (resourcelog_init(&ranklog, outlogpath, RESOURCE_READ_LOG, rman->nslist[response->request.nsindex])) {
-            fprintf(stderr, "ERROR: Failed to open the output log of Rank %zu for NS \"%s\": \"%s\"\n",
-                     ranknum, rman->nslist[response->request.nsindex]->idstr, outlogpath);
-            free(outlogpath);
-            rman->fatalerror = 1;
-            return -1;
-         }
-         // process the work log
-         if (response->errorlog) {
-            // note a non-fatal error (non-zero exit code for this run)
-            rman->nonfatalerror = 1;
-            // parse through our output path, looking for the final path element
-            char* parselogpath = outlogpath;
-            char* finelem = outlogpath;
-            while (*parselogpath != '\0') {
-               // check for '/' separator and skip beyond it
-               if (*parselogpath == '/') { while (*parselogpath == '/') { parselogpath++; finelem = parselogpath; } }
-               else { parselogpath++; } // only increment here if we aren't already doing so above (avoids possible OOB error with '/' final char)
-            }
-            // identify the errorlog output location
-            int logdirlen = finelem - outlogpath; // pointer arithmetic to get prefix dir string length
-            int fnamelen = parselogpath - finelem;
-            size_t errloglen = logdirlen + strlen(ERROR_LOG_PREFIX) + fnamelen;
-            char* errlogpath = malloc(sizeof(char) * (errloglen + 1));
-            if (snprintf(errlogpath, errloglen + 1, "%.*s%s%s", logdirlen, outlogpath, ERROR_LOG_PREFIX, finelem) != errloglen) {
-               fprintf(stderr, "ERROR: Failed to generate the error log path of Rank %zu for NS \"%s\"\n",
-                        ranknum, rman->nslist[response->request.nsindex]->idstr);
-               resourcelog_term(&ranklog, NULL, 0);
-               free(outlogpath);
-               rman->fatalerror = 1;
-               return -1;
-            }
-            // open the error log for write
-            RESOURCELOG errlog = NULL;
-            if (resourcelog_init(&errlog, errlogpath, RESOURCE_RECORD_LOG, rman->nslist[response->request.nsindex])) {
-               fprintf(stderr, "ERROR: Failed to open the error log of Rank %zu: \"%s\"\n", ranknum, errlogpath);
-               resourcelog_term(&ranklog, NULL, 0);
-               free(errlogpath);
-               free(outlogpath);
-               rman->fatalerror = 1;
-               return -1;
-            }
-            // replay the logfile into the error log location
-            if (resourcelog_replay(&ranklog, &errlog, error_only_filter)) {
-               fprintf(stderr, "ERROR: Failed to replay errors from logfile \"%s\" into \"%s\"\n", outlogpath, errlogpath);
-               resourcelog_term(&errlog, NULL, 1);
-               resourcelog_term(&ranklog, NULL, 0);
-               free(errlogpath);
-               free(outlogpath);
-               rman->fatalerror = 1;
-               return -1;
-            }
-            if (resourcelog_term(&errlog, NULL, 0)) {
-               fprintf(stderr, "ERROR: Failed to finalize error logfile: \"%s\"\n", errlogpath);
-               free(errlogpath);
-               free(outlogpath);
-               rman->fatalerror = 1;
-               return -1;
-            }
-            free(errlogpath);
-         }
-         else {
-            // simply delete the output logfile of this rank
-            if (resourcelog_term(&ranklog, NULL, 1)) {
-               fprintf(stderr, "ERROR: Failed to delete the output log of Rank %zu for NS \"%s\": \"%s\"\n",
-                        ranknum, rman->nslist[response->request.nsindex]->idstr, outlogpath);
-               free(outlogpath);
-               rman->fatalerror = 1;
-               return -1;
-            }
-         }
-         free(outlogpath);
-      } // end of response content processing
-      // this rank needs work to process, with no preference for NS
-      if (rman->oldlogs) {
-         // start by checking for old resource logs to process
-         if (hash_reset(rman->oldlogs) < 0) {
-            fprintf(stderr, "ERROR: Failure of hash_reset() when trying to identify logfiles for rank %zu to process\n",
-                             ranknum);
-            rman->fatalerror = 1;
-            return -1;
-         }
-         HASH_NODE* resnode = NULL;
-         while (resnode) {
-            // check for any nodes with log processing requests remaining
-            loginfo* linfo = (loginfo*)(resnode->content);
-            if (linfo->logcount) {
-               // use the request at the tail of our list
-               LOG(LOG_INFO, "Passing out log %zu for NS \"%s\" from iteration \"%s\"\n",
-                    linfo->logcount,
-                    rman->nslist[linfo->nsindex]->idstr,
-                    (linfo->requests + (linfo->logcount - 1))->iteration);
-               *request = *(linfo->requests + (linfo->logcount - 1));
-               linfo->logcount--;
-               return 1;
-            }
-            if (hash_iterate(rman->oldlogs, &resnode) < 0) {
-               fprintf(stderr,
-                        "ERROR: Failure of hash_iterate() when trying to identify logfiles for rank %zu to process\n",
-                        ranknum);
-               rman->fatalerror = 1;
-               return -1;
-            }
-         }
-         printf("  -- All old logfiles have been handed out for processing --\n");
-         // it seems we've iterated over our entire oldlogs table, with no requests remaining
-         // need to free our table
-         size_t ncount = 0;
-         if (hash_term(rman->oldlogs, &resnode, &ncount)) {
-            fprintf(stderr, "ERROR: Failure of hash_term() after passing out all logfiles for processing\n");
-            rman->fatalerror = 1;
-            return -1;
-         }
-         rman->oldlogs = NULL;
-         // free all subnodes and requests
-         size_t nindex = 0;
-         for (; nindex < ncount; nindex++) {
-            if ((resnode+nindex)->name) { free((resnode+nindex)->name); }
-            loginfo* linfo = (loginfo*)((resnode+nindex)->content);
-            if (linfo) {
-               if (linfo->requests) { free(linfo->requests); }
-               free(linfo);
-            }
-         }
-         free(resnode); // these were allocated in one block, and thus require only one free()
-      }
-      // no resource logs remain to process
-      if (rman->execprevroot) {
-         // if we are picking up a previous run, this means no more work remains at all
-         LOG(LOG_INFO, "Signaling Rank %zu to terminate, as no resource logs remain\n", ranknum);
-         request->type = TERMINATE_WORK;
-         return 1;
-      }
-      // first, check specifically for NSs that have yet to be processed at all
-      size_t nsindex = 0;
-      char anynswork = 0;
-      for (; nsindex < rman->nscount; nsindex++) {
-         if (rman->distributed[nsindex] == 0) {
-            // this NS still has yet to be worked on at all
-            request->type = NS_WORK;
-            request->nsindex = nsindex;
-            request->refdist = rman->distributed[nsindex];
-            request->iteration[0] = '\0';
-            request->ranknum = ranknum;
-            rman->distributed[nsindex]++; // note newly distributed range
-            LOG(LOG_INFO, "Passing out reference range %zu of NS \"%s\" to Rank %zu\n",
-                 request->refdist, rman->nslist[nsindex]->idstr, ranknum);
-            printf("  Rank %zu is beginning work on NS \"%s\" (ref range %zu)\n",
-                    ranknum, rman->nslist[nsindex]->idstr, request->refdist);
-            // check through remaining namespaces for any undistributed work
-            for (; anynswork == 0  &&  nsindex < rman->nscount; nsindex++) {
-               if (rman->distributed[nsindex] < rman->workingranks) { break; }
-            }
-            if (nsindex == rman->nscount) {
-               // just handed out the last NS ref range for processing
-               printf("  -- All NS reference ranges have been handed out for processing --\n");
-            }
-            return 1;
-         }
-         else if (rman->distributed[nsindex] < rman->workingranks) { anynswork = 1; }
-      }
-      // next, check for NSs with ANY remaining work to distribute
-      nsindex = 0;
-      for (; nsindex < rman->nscount; nsindex++) {
-         if (rman->distributed[nsindex] < rman->workingranks) {
-            // this NS still has reference ranges to be scanned
-            request->type = NS_WORK;
-            request->nsindex = nsindex;
-            request->refdist = rman->distributed[nsindex];
-            request->iteration[0] = '\0';
-            request->ranknum = ranknum;
-            rman->distributed[nsindex]++; // note newly distributed range
-            LOG(LOG_INFO, "Passing out reference range %zu of NS \"%s\" to Rank %zu\n",
-                 request->refdist, rman->nslist[nsindex]->idstr, ranknum);
-            printf("  Rank %zu is picking up work on NS \"%s\" (ref range %zu)\n",
-                    ranknum, rman->nslist[nsindex]->idstr, request->refdist);
-            // check through remaining namespaces for any undistributed work
-            for (; nsindex < rman->nscount; nsindex++) {
-               if (rman->distributed[nsindex] < rman->workingranks) { break; }
-            }
-            if (nsindex == rman->nscount) {
-               // just handed out the last NS ref range for processing
-               printf("  -- All NS reference ranges have been handed out for processing --\n");
-            }
-            return 1;
-         }
-      }
-      // no NS reference ranges remain to process, so we now need to signal termination
-      LOG(LOG_INFO, "Signaling Rank %zu to terminate, as no NS reference ranges remain\n", ranknum);
-      request->type = TERMINATE_WORK;
-      return 1;
-   }
-   else if (response->request.type == NS_WORK  ||  response->request.type == RLOG_WORK) {
-      // this rank needs work to process, specifically in the same NS
-      if (rman->oldlogs) {
-         // start by checking for old resource logs to process
-         HASH_NODE* resnode = NULL;
-         if (hash_lookup(rman->oldlogs, rman->nslist[response->request.nsindex]->idstr, &resnode)) {
-            fprintf(stderr, "ERROR: Failure of hash_lookup() when looking for logfiles in NS \"%s\" for rank %zu\n",
-                             rman->nslist[response->request.nsindex]->idstr, ranknum);
-            rman->fatalerror = 1;
-            return -1;
-         }
-         // check for any nodes with log processing requests remaining
-         loginfo* linfo = (loginfo*)(resnode->content);
-         if (linfo->logcount) {
-            // use the request at the tail of our list
-            LOG(LOG_INFO, "Passing out log %zu for NS \"%s\" from iteration \"%s\"\n",
-                 linfo->logcount,
-                 rman->nslist[linfo->nsindex]->idstr,
-                 (linfo->requests + (linfo->logcount - 1))->iteration);
-            *request = *(linfo->requests + (linfo->logcount - 1));
-            linfo->logcount--;
-            return 1;
-         }
-      }
-      // no resource logs remain to process in the active NS of this rank
-      if (rman->execprevroot) {
-         // if we are picking up a previous run, this means no more work remains for the active NS at all
-         LOG(LOG_INFO, "Signaling Rank %zu to complete and quiesce, as no resource logs remain in NS \"%s\"\n",
-              ranknum, rman->nslist[response->request.nsindex]->idstr);
-         *request = response->request; // populate request with active NS values
-         request->type = COMPLETE_WORK;
-         return 1;
-      }
-      // check for any remaining work in the rank's active NS
-      if (rman->distributed[response->request.nsindex] < rman->workingranks) {
-         request->type = NS_WORK;
-         request->nsindex = response->request.nsindex;
-         request->refdist = rman->distributed[response->request.nsindex];
-         request->iteration[0] = '\0';
-         request->ranknum = ranknum;
-         rman->distributed[response->request.nsindex]++; // note newly distributed range
-         LOG(LOG_INFO, "Passing out reference range %zu of NS \"%s\" to Rank %zu\n",
-              request->refdist, rman->nslist[response->request.nsindex]->idstr, ranknum);
-         // check through remaining namespaces for any undistributed work
-         size_t nsindex = 0;
-         for (; nsindex < rman->nscount; nsindex++) {
-            if (rman->distributed[nsindex] < rman->workingranks) { break; }
-         }
-         if (nsindex == rman->nscount) {
-            // just handed out the last NS ref range for processing
-            printf("  -- All NS reference ranges have been handed out for processing --\n");
-         }
-         return 1;
-      }
-      // all work in the active NS has been completed
-      LOG(LOG_INFO, "Signaling Rank %zu to complete and quiesce, as no reference ranges remain in NS \"%s\"\n",
-           ranknum, rman->nslist[response->request.nsindex]->idstr);
-      *request = response->request; // populate request with active NS values
-      request->type = COMPLETE_WORK;
-      return 1;
-   }
-
-   LOG(LOG_ERR, "Encountered unrecognized response type\n");
-   fprintf(stderr, "ERROR: Encountered response to an unknown request type\n");
-   rman->fatalerror = 1;
-   return -1;
-}
-
 
 //   -------------   CORE BEHAVIOR LOOPS   -------------
 
@@ -2088,13 +720,23 @@ static int parse_args(int argc, char** argv,
    }
 
    // substitute in appropriate threshold values, if specified
-   if (rman->gstate.thresh.gcthreshold) { rman->gstate.thresh.gcthreshold = args->thresh.gc; }
-   if (rman->gstate.thresh.rebuildthreshold) {
-      if (rman->gstate.lbrebuild) { rman->gstate.thresh.rebuildthreshold = args->thresh.rbl; }
-      else { rman->gstate.thresh.rebuildthreshold = args->thresh.rbm; }
+   if (rman->gstate.thresh.gcthreshold) {
+       rman->gstate.thresh.gcthreshold = args->thresh.gc;
    }
-   if (rman->gstate.thresh.repackthreshold) { rman->gstate.thresh.repackthreshold = args->thresh.rp; }
-   if (rman->gstate.thresh.cleanupthreshold) { rman->gstate.thresh.cleanupthreshold = args->thresh.cl; }
+   if (rman->gstate.thresh.rebuildthreshold) {
+      if (rman->gstate.lbrebuild) {
+          rman->gstate.thresh.rebuildthreshold = args->thresh.rbl;
+      }
+      else {
+          rman->gstate.thresh.rebuildthreshold = args->thresh.rbm;
+      }
+   }
+   if (rman->gstate.thresh.repackthreshold) {
+       rman->gstate.thresh.repackthreshold = args->thresh.rp;
+   }
+   if (rman->gstate.thresh.cleanupthreshold) {
+       rman->gstate.thresh.cleanupthreshold = args->thresh.cl;
+   }
 
    return 0;
 }
@@ -2109,7 +751,6 @@ static int parse_args(int argc, char** argv,
     cleanup_abort((rman), (erasurelock))
 
 //   -------------    STARTUP BEHAVIOR     -------------
-
 
 int main(int argc, char** argv) {
 #ifdef RMAN_USE_MPI
@@ -2193,185 +834,31 @@ int main(int argc, char** argv) {
       rmanstate_fini(&rman, 0);
       RMAN_ABORT();
    }
+
    if ((rman.config = config_init(args.config_path, &erasurelock)) == NULL) {
-      fprintf(stderr, "ERROR: Failed to initialize MarFS config: \"%s\"\n", args.config_path);
-      cleanup_abort(rman, erasurelock);
+       print_cleanup_abort(rman, erasurelock,
+                           "ERROR: Failed to initialize MarFS config: \"%s\"\n", args.config_path);
    }
 
-   // Identify our target NS
-   marfs_position pos = {
-      .ns = NULL,
-      .depth = 0,
-      .ctxt = NULL
-   };
-   if (config_establishposition(&pos, rman.config)) {
-      print_cleanup_abort(rman, erasurelock,
-                          "ERROR: Failed to establish a MarFS root NS position\n");
+   if (find_namespaces(&rman, args.ns_path, args.recurse) != 0) {
+       print_cleanup_abort(rman, erasurelock,
+                           "Error: Failed to find namespace\n");
    }
-   char* nspathdup = strdup(args.ns_path);
-   if (nspathdup == NULL) {
-      print_cleanup_abort(rman, erasurelock,
-                          "ERROR: Failed to duplicate NS path string: \"%s\"\n", args.ns_path);
-   }
-   int travret = config_traverse(rman.config, &pos, &nspathdup, 1);
-   if (travret < 0) {
-      if (rman.ranknum == 0)
-         fprintf(stderr, "ERROR: Failed to identify NS path target: \"%s\"\n", args.ns_path);
-      free(nspathdup);
-      cleanup_abort(rman, erasurelock);
-   }
-   if (travret) {
-      if (rman.ranknum == 0)
-         fprintf(stderr, "ERROR: Path target is not a NS, but a subpath of depth %d: \"%s\"\n", travret, args.ns_path);
-      free(nspathdup);
-      cleanup_abort(rman, erasurelock);
-   }
-   free(nspathdup);
-   // Generate our NS list
-   marfs_ns* curns = pos.ns;
-   rman.nscount = 1;
-   rman.nslist = malloc(sizeof(marfs_ns*));
-   *(rman.nslist) = pos.ns;
-   while (curns) {
-      // we can use hash_iterate, as this is guaranteed to be the only proc using this config struct
-      HASH_NODE* subnode = NULL;
-      int iterres = 0;
-      if (curns->subspaces  &&  args.recurse) {
-         iterres = hash_iterate(curns->subspaces, &subnode);
-         if (iterres < 0) {
-            print_cleanup_abort(rman, erasurelock,
-                                "ERROR: Failed to iterate through subspaces of \"%s\"\n", curns->idstr);
-         }
-         else if (iterres) {
-            marfs_ns* subspace = (marfs_ns*)(subnode->content);
-            // only process non-ghost subspaces
-            if (subspace->ghtarget == NULL) {
-               LOG(LOG_INFO, "Adding subspace \"%s\" to our NS list\n", subspace->idstr);
-               // note and enter the subspace
-               rman.nscount++;
-               // yeah, this is very inefficient; but we're only expecting 10s to 1000s of elements
-               marfs_ns** newlist = realloc(rman.nslist, sizeof(marfs_ns*) * rman.nscount);
-               if (newlist == NULL) {
-                  print_cleanup_abort(rman, erasurelock,
-                                      "ERROR: Failed to allocate NS list of length %zu\n", rman.nscount);
-               }
-               rman.nslist = newlist;
-               *(rman.nslist + rman.nscount - 1) = subspace;
-               curns = subspace;
-               continue;
-            }
-         }
-      }
-      if (iterres == 0) {
-         // check for completion condition
-         if (curns == pos.ns) {
-            // iteration over the original NS target means we're done
-            curns = NULL;
-         }
-         else {
-            curns = curns->pnamespace;
-         }
-      }
-   }
-   // abandon our current position
-   if (config_abandonposition(&pos)) {
-      print_cleanup_abort(rman, erasurelock,
-                          "WARNING: Failed to abandon MarFS traversal position\n");
-   }
+
    // complete allocation of required state elements
    rman.distributed = calloc(sizeof(size_t), rman.nscount);
    rman.terminatedworkers = calloc(sizeof(char), rman.totalranks);
    rman.walkreport = calloc(sizeof(*rman.walkreport), rman.nscount);
    rman.logsummary = calloc(sizeof(*rman.logsummary), rman.nscount);
 
-   // check for previous run execution
-   if (rman.execprevroot) {
-      // open the summary log of that run
-      size_t alloclen = strlen(rman.execprevroot) + 1 + strlen(rman.iteration) + 1 + strlen(SUMMARY_FILENAME) + 1;
-      char* sumlogpath = malloc(sizeof(char) * alloclen);
-      snprintf(sumlogpath, alloclen, "%s/%s/%s", rman.execprevroot, rman.iteration, SUMMARY_FILENAME);
-      FILE* sumlog = fopen(sumlogpath, "r");
-      if (sumlog == NULL) {
-         const int err = errno;
-         print_cleanup_abort(rman, erasurelock,
-                             "ERROR: Failed to open previous run's summary log: \"%s\" (%s)\n",
-                             sumlogpath, strerror(err));
-      }
-      if (parse_program_args(&rman, sumlog)) {
-         fclose(sumlog);
-         print_cleanup_abort(rman, erasurelock,
-                             "ERROR: Failed to parse previous run info from summary log: \"%s\"\n",
-                             sumlogpath);
-      }
-      if (rman.quotas) {
-         if (rman.ranknum == 0)
-            fprintf(stderr, "WARNING: Ignoring quota processing for execution of previous run\n");
-         rman.quotas = 0;
-      }
-      fclose(sumlog);
-      free(sumlogpath);
-      if (rman.gstate.dryrun == 0) {
-         if (rman.ranknum == 0)
-            fprintf(stderr, "ERROR: Cannot pick up execution of a non-'dry-run'\n");
-         cleanup_abort(rman, erasurelock);
-      }
-      // incorporate logfiles from the previous run
-      if (rman.ranknum == 0  &&  findoldlogs(&rman, rman.execprevroot, 0)) {
-         print_cleanup_abort(rman, erasurelock,
-                             "ERROR: Failed to identify previous run's logfiles: \"%s\"\n",
-                             rman.execprevroot);
-      }
-   }
-   else if (rman.gstate.dryrun == 0  &&  rman.ranknum == 0) {
-      // otherwise, scan for and incorporate logs from previous modification runs
-      if (findoldlogs(&rman, rman.logroot, args.thresh.skip)) {
-         print_cleanup_abort(rman, erasurelock,
-                             "ERROR: Failed to scan for previous iteration logs under \"%s\"\n",
-                             rman.logroot);
-      }
+   if (read_last_log(&rman, args.thresh.skip) != 0) {
+       print_cleanup_abort(rman, erasurelock,
+                           "ERROR: Failed to open previous run's log\n");
    }
 
-   // rank zero needs to output our summary header
-   if (rman.ranknum == 0) {
-      // open our summary log
-      size_t alloclen = strlen(rman.logroot) + 1 + strlen(rman.iteration) + 1 + strlen(SUMMARY_FILENAME) + 1;
-      char* sumlogpath = malloc(sizeof(char) * alloclen);
-      size_t printres = snprintf(sumlogpath, alloclen, "%s/%s", rman.logroot, rman.iteration);
-      if (mkdir(sumlogpath, 0700)  &&  errno != EEXIST) {
-         print_cleanup_abort(rman, erasurelock,
-                             "ERROR: Failed to create summary log parent dir: \"%s\"\n",
-                             sumlogpath);
-      }
-      printres += snprintf(sumlogpath + printres, alloclen - printres, "/" SUMMARY_FILENAME);
-      int sumlog = open(sumlogpath, O_WRONLY | O_CREAT | O_EXCL, 0700);
-      if (sumlog < 0) {
-         print_cleanup_abort(rman, erasurelock,
-                             "ERROR: Failed to open summary logfile: \"%s\"\n",
-                             sumlogpath);
-      }
-      rman.summarylog = fdopen(sumlog, "w");
-      if (rman.summarylog == NULL) {
-         print_cleanup_abort(rman, erasurelock,
-                             "ERROR: Failed to convert summary logfile to file stream: \"%s\"\n",
-                             sumlogpath);
-      }
-      // output our program arguments to the summary file
-      if (output_program_args(&rman)) {
-         print_cleanup_abort(rman, erasurelock,
-                             "ERROR: Failed to output program arguments to summary log: \"%s\"\n",
-                             sumlogpath);
-      }
-      free(sumlogpath);
-
-      // print out run info
-      printf("Processing %zu Total Namespaces (%sTarget NS \"%s\")\n",
-              rman.nscount, (args.recurse) ? "Recursing Below " : "", (rman.nslist[0])->idstr);
-      printf("   Operation Summary:%s%s%s%s%s\n",
-              (rman.gstate.dryrun) ? " DRY-RUN" : "", (rman.quotas) ? " QUOTAS" : "",
-              (rman.gstate.thresh.gcthreshold) ? " GC" : "",
-              (rman.gstate.thresh.repackthreshold) ? " REPACK" : "",
-              (rman.gstate.thresh.rebuildthreshold) ?
-               ((rman.gstate.lbrebuild) ? " REBUILD(LOCATION)" : " REBUILD(MARKER)") : "");
+   if (summary_log_setup(&rman, args.recurse) != 0) {
+      print_cleanup_abort(rman, erasurelock,
+                          "ERROR: Failed to set up summary\n");
    }
 
 #ifdef RMAN_USE_MPI
