+++ src/rsrc_mgr/resourcethreads.c
@@ -57,17 +57,8 @@ https://github.com/jti-lanl/aws4c.
 GNU licenses can be found at http://www.gnu.org/licenses/.
 */
 
-#include "marfs_auto_config.h"
-#ifdef DEBUG_RM
-#define DEBUG DEBUG_RM
-#elif (defined DEBUG_ALL)
-#define DEBUG DEBUG_ALL
-#endif
-#define LOG_PREFIX "resourcethreads"
-#include "logging/logging.h"
-
-#include "resourcethreads.h"
-
+#include "rsrc_mgr/consts.h"
+#include "rsrc_mgr/resourcethreads.h"
 
 //   -------------   RESOURCE INPUT FUNCTIONS    -------------
